Add a Bootstrap.registerAsynchronously method
Motivation:
In #8516 it was proposed to at some point remove the specialised ChannelFuture and ChannelPromise.
Or at least make them not extend Future and Promise, respectively.
One pain point encountered in this discussion is the need to get access to the channel object after it has been initialised, but without waiting for the channel registration to propagate through the pipeline.

Modification:
Add a Bootstrap.registerAsynchronously method, which will return a Future<Channel>.
This future will provide access to the channel as soon as it has been initialised.
The registration of the channel through the pipeline will only start after the promise associated with the returned future has been completed.

Result:
It's now possible to obtain a channel initialised from a bootstrap, without using ChannelFuture, and without waiting for channel registration to propagate through the pipeline.