Initial draft of cleaned up Future/Promises API
Motivation:
The generics for the existing futures, promises, and listeners are too complicated.
Some of this complication comes from the existence of ChannelPromise and ChannelFuture.

Modification:
This shows a path forward where Future and Promise are "leaf" types, which means no code should need to generalise over the type of futures.
Because of this, GenericFutureListener is now deprecated and FutureListener should be used instead.
To get away from ChannelFuture, a new listener type is introduced which take an additional context argument, which is passed along when the listener is added to a future.

Result:
Cleaner generics and hopefully more straight forward code.

This is an intermediate change, which does not yet compile.
This is just to show the idea.