Fix CRLF split between 2 buffers before separator
Fix issue #11668

When a buffer ends with the last chunk for a part but containing
only a CR and no LF neither separator, it ends up with the CR
included within the content of the part, which is wrong.

Rare case, but fixed by checking when not getting the delimiter,
not finding a LineBreak, we check if the last byte of the chunk is a CR
and if the content could be added to the current part (current length +
current buffer length - 1) could lead to the end of the part.

If so, then the CR can be safely ignored, waiting for the future LF and
separator, and the buffer can be added to the part content.

If not, the CR is part of the content and shall be included.

Add a non regression test.