Add a Resource.touch method
Motivation:
It is useful for debugging to be able to add important "touch points" to the life cycle trace of an object, so that life cycle issues can be debugged with higher fidelity information than what the usual acquire/close/send/receive trace points provide.

Modification:
A `touch` method is added to the `Resource` interface, and `ResourceSupport` provides an implementation.
Implementations have also been added to `BufferHolder`, `BufferStub`, `AdaptableBuffer` (for compatibility with `ByteBuf` API), and `DefaultCompositeBuffer`.
The LifecycleTracer has also been extended to support this new trace point, and the internals have been streamlined a bit.
No test has been added at this point, because the lifecycle tracer API needs more work before it is testable.
This will come in a later PR.

Result:
It is now possible to "touch" buffers to add more information to their life cycle traces.