Fix little-endian get/set methods in SwappedByteBuf
Motivation:
The `*LE` family of accessor methods are documented to always perform accesses in little-endian byte order, regardless of what `ByteBuf.order()` has been set on a buffer.
This was not always honoured by the `SwappedByteBuf`, and many methods were fixed in #10747, except the get and set methods were missed by that PR.

Modification:
Fix the methods that were missed in #10747 and make the testing more thorough by also covering get and set methods.

Result:
Any `ByteBuf` accessor method that is suffixed with `LE` now always perform a little-endian access, regardless of the buffers configured byte order.

This fixes #11700