Fix CleanerDrop and how it is used
Motivation:
When porting the CleanerDrop over to the main Netty repo, it got broken and ended up accidentally holding on to the cleaner-monitored object from the cleaner action.
Such a reference loop means the cleaner will never clean any buffers that leak.

Modification:
Break the reference cycle in the CleanerDrop and ensure that recoverable memory is actually recovered if it leaks.
We test for this by instrumenting the MemoryManager, because these will be indirectly told about leaks.

Result:
Buffers that leak are now properly deallocated or returned to their originating pool.