Avoid allocating new arenas per thread when caching is disabled
Motivation:
Even when thread-local caching is disabled, the PooledBufferAllocator still needs to return a PoolThreadCache instance that contain an arena from the pool.
This is how allocation calls generally access arenas in the pool, regardless of thread-local caching.
This means if the PoolThreadCache has no arena, then the allocation will fall back to the unpooled path.
In other words, if thread-local caching failed, then _all_ pooling and caching would be disabled for the allocator.
This is clearly suboptimal behaviour.

Modification:
The PoolThreadCache now acquires a reference to the least-used arena at the time of initialisation.
Since buffers and leak detection rely on cleaners and phantom references to the memory, the PoolThreadCache will have to only hold a weak reference to the arena and its memory regions.
Otherwise, the thread-local caches would hold strong references to the memory, and prevent it from being reclaimed.
The arenas should only be held live by the central pool for caching, or by buffers for use.

Result:
The new Buffers are pooled and cached more effectively, when the allocating thread does not support FastThreadLocals, or when caching is turned off.