Add ByteBufAdaptor.intoByteBuf method
Motivation:
During the transition period, it will be convenient to have a utility method for converting a Buffer into a ByteBuf in the most efficient way.
A ByteBufConvertible interface already exists for this purpose, but not all Buffer instances are guarenteed to implement this, and those that did had been broken.

Modification:
Fix the broken ByteBufConvertible implementations.
Add a ByteBufAdaptor.intoByteBuf method that uses ByteBufConvertible when available, and a direct ByteBufAdaptor wrapping when not.

Result:
Converting a Buffer to a ByteBuf is now just one method call away.