Re-enable ByteBuf test for ByteBufAdaptor
Motivation:
We should run as many ByteBuf tests on the ByteBufAdaptor as possible.
This test had previously been disabled because at the time it was not possible to allocate zero-sized buffers using the new API.

Modification:
Re-enable ByteBufAdaptorTest that was relying on zero-sized buffer allocations.

Result:
Our ByteBufAdaptor is more thoroughly tested.