Make implementing Resource.touch optional
Motivation:
The Resource.touch method is only used for debugging life-cycle issues, and is then only useful for objects that track their life-cycles.
Buffers do this, but there will be many other implementations of the Resource interface where it makes less sense.

Modification:
Add a default implementation of the Resource.touch method that just returns the resource instance.
This is enough to conform to the specification of the method as declared.

Result:
The touch method declaration no longer pollutes implementations where it serves no purpose.
This eases the burden of implementing the Resource interface.