The ByteBufAdaptor should cache and reuse its ByteBufAllocatorAdaptor instance
Motivation:
It can be costly to create BufferAllocators all the time, so we should avoid doing that in ByteBufAdaptor.intoByteBuf.

Modification:
Create a globally shared and cached ByteBufAllocatorAdaptor instance, which is based on the DefaultGlobalBufferAllocator.
We also create a counter-part allocator, because the allocator adaptor needs both an on-heap and an off-heap allocator.
Both of these allocators are kept as constants and initialised once.
A new UncloseableBufferAllocator class is added to be the base class for both the global allocator and its counter-part.

Result:
No more allocators created when a Buffer is converted into a ByteBuf.