Epoll compiled on GLIBC only supports GLIBC at runtime
Motivation:
It is possible to run into JVM crashes when our native binaries are compiled for one libc but then linked with another at runtime.
There is no guarantee that their ABIs will be compatible, and this can cause all sorts of trouble.
All our Linux binaries are compiled for GLIBC, so it makes sense to check that any binary compiled for GLIBC also ends up with GLIBC at runtime.
Especially since alternative libc implementations, such as musl, are getting increasingly popular in a containerised/serverless world.

Modification:
Add a check when loading the epoll native library that verifies: if the library was compiled for GLIBC, then we must also have a GLIBC loaded at runtime.
The key is to make this check without making any libc calls in our native code.
We make this check by reading `/proc/self/maps`, which contain paths for all loaded libraries in our process.
GLIBC tends to have paths with a particular pattern to their name, so we look for that.
At compile time, we rely on `__GLIBC__` to determine if the binary is compiled for GLIBC or not.
We cannot check for Musl because they intentionally expose no equivalent define to check for.
By checking for GLIBC specifically, we also end up handling all other alternative libc implementations, and not just Musl.

Result:
Epoll is now reported as not available when running a GLIBC binary on a Musl - or any other alternative libc - based system.
Previously, epoll would be reported as available, but could cause JVM crashes.