Add server-side TCP FastOpen support for macOS
Motivation:
This completes the story of TCP FastOpen support for macOS.

Modification:
On macOS, to enable TCP FastOpen for a listening socket, we need to set it as a socket option after the bind(2) and listen(2) calls (as opposed to before the listen call, like on Linux, or at config time).
Native methods are added to BsdSocket for getting and setting the TCP_FASTOPEN socket option.
Since activation happens at bind-time, a field is also added to KQueueServerChannelConfig, to track if TCP FastOpen should be enabled or not.
By default, it is off, which leaves this socket option unset to the operating system.

Result:
We now have both client and server support for TCP FastOpen on macOS.