Make the needTask flag lowering consistent in ReferenceCountedOpenSslEngine
Motivation:
The needTask flag needs to be lowered before an off-loaded task is executed.
This is because the executed task might want to inspect the needTask state, and it should not see itself in that case.

Modification:
Lower the needTask state before the task execution, not only for asynchronous tasks but also for synchronous tasks.

Result:
More consistent behaviour.