First draft of Buffer.firstOffsetOf
Motivation:
Being able to find a particular byte within a buffer is important and performance critical in a number of use cases.

Modification:
Add a Buffer.firstOffsetOf method to Buffer.
The implementation is based on the existing ByteBufUtil.firstIndexOf.
Tests and benchmarks are also added

Result:
We now have a fast way to find the first occurrence of a particular byte in a buffer.
The performance of this method is in the ballpark of the existing ByteBufUtil.firstIndexOf; some cases are better, others are worse.