New leak detector for the new Buffer APIs (#11725)
Motivation:
Our buffers are reference counted, which means it's possible to leak them.
Fortunately, each is also attached to a Cleaner, which means leaked memory is automatically reclaimed, though perhaps with some delay.
It is useful to be able to be notified when a buffer leaks, so these can be fixed and the code can run with a tighter memory budget.

Modification:
Add a `MemoryManager.onLeakDetected()` method which takes a callback, that will be given a `LeakInfo` object when a leak is detected.
The method returns a `SafeCloseable` which, when closed, will uninstall the callback.
It is possible to install the same callback multiple times, but it will only be invoked once per leak.
The leak info can be inspected to recover the life cycle of the object that leaked.
This makes use of the `LifecycleTracer` mechanism that already exists for objects that extend `ResourceSupport` - primarily `Buffer` objects.
The leak detection mechanism is controlled by a new `LeakDetection` class, which is internal to the new buffer API.
The leak detection gets notified about leaks by the `CleanerDrop`, which is currently Buffer specific.
In principle, other mechanisms can be devised to report leaks for other objects.
Leak detection and lifecycle tracing is off by default, but are automatically turned on when the first on-leak callback is installed.
To get stack traces in the reported leaks, the life-cycle tracer needs to be configured to enable this separately.
A `LoggingLeakCallback` class is added, which produces output similar to the older `ResourceLeakDetector`.
In the process of implementing this, and adding tests for it, a number of bugs related to `CleanerDrop` and pooling allocators, etc. have been fixed.
Now that we have a working leak detector, we are able to prove that the cleaner-based recovery mechanism actually works.

Result:
Leak detection for the new Buffer API, and correct cleaner-based recovery of leaked buffers.