Make it possible to query Epoll if TCP FastOpen is available
Motivation:
While TCP FastOpen is mostly a pure optimisation (albeit one that demands idempotency of TFO messages), it can still sometimes be useful to query whether it can be expected to be available.
This information can for instance be useful for telemetry, where you want to include if TFO is available to the system, to see if it helps overall.
For epoll, we have such fields already, but they are not public.
The public field we do have, is deprecated and conflates clients and servers.

Modification:
Add static methods to Epoll that delegates to the existing IS_SUPPORTING_TCP_FASTOPEN_{CLIENT/SERVER} fields.
Whether TFO is available is also predicated upon epoll itself being available.

Result:
People, who specifically rely on epoll, can now query their system in pure Java about whether TFO is available or not.