Add a MemoryManager.unsafeWrap method
Motivation:
There are rare use cases where we might wish to directly wrap a byte array in a Buffer.
This is unsafe because the memory, the byte array, can be aliased.
Hence, this functionality should not be in plain sight.
These buffers also cannot be pooled, so there is no reason to add this functionality to the allocator interface.
The MemoryManager interface is a better location.
These buffers must also be read-only.

Modification:
Add a MemoryManager.unsafeWrap method, that takes a byte array and uses the default memory manager to create an on-heap buffer wrapping the byte array.
A special AllocationType has been added for this purpose.
This work also uncovered a few bugs and inconsistencies around read-only buffers.
These have been fixed, and test coverage has increased.

Result:
It is now possible to directly wrap a byte array in a read-only Buffer.