Update Groovy versions
Motivation:
The old Groovy versions are running into compatibility issues with the latest JDK snapshots.

Modification:
Update the groovy maven plugin and its groovy dependency to the latest versions.
It looks like groovy-all is now a pom-dependency, so we switch to just groovy since we don't need what's in -all beyond that.

Result:
We now build on the latest JDK snapshots.