Make CharSeqValueConverter.convertToMillis throw DateTimeParseException
Motivation:
The `convertToTimeMillis` method was using `PlatformDependent.throwException` to throw a checked exception that the method did not declare.
We should avoid doing that, and use proper exceptions instead.

Modification:
Make `convertToTimeMillis` throw the runtime `DateTimeParseException` instead.
As a kind of `RuntimeException`, we do not need to declare this exception, yet we can throw it.
Also update and add to the test coverage of this method.
Direct usages of this method have also been updated and more thoroughly tested.

Result:
One fewer places where we throw undeclared checked exceptions.