Rewrite and simplify Recycler
Motivation:
We have in the past gotten reports of weird errors that we haven't been able to reproduce, but which could be explained by bugs in the Recycler.
The reason is that the Recycler was mixing concurrency, and interactions with the GC, in ways that were difficult to reason about.
It is unwise to leave suspicous-looking code unaddressed.

Modification:
The Recycler has been rewritten to use a much simpler algorithm, and now completely avoids the use of weak references.

The WeakOrderQueue is gone, and there is no longer a concept of a shared max capacity - only thread-local capacity.

There is also no more scavenging.
It turns out that other code was relying on the ability to forget references to objects they acquired from a Recycler, and this is hard to support if there is a shared pool for moving handles between threads.

Each thread now has a local pool (an ArrayDeque), and an external inbox (also an ArrayDeque, but access is synchronised).
Calls to get() and recycle() in the same thread will just operate on the local pool like a stack.
Calls to recycle() from a thread that did not get() that object (the object was acquired in another thread) will place the object in the originating pools inbox.
When a thread then goes to get() another object, they will first check if they have anything in their inbox, and if so, drain it into their local pool.

Result:
The Recycler code is much easier to understand.