Rewrite and simplify Recycler (#11858)
Motivation:
We have in the past gotten reports of weird errors that we haven't been able to reproduce, but which could be explained by bugs in the Recycler.
The reason is that the Recycler was mixing concurrency, and interactions with the GC, in ways that were difficult to reason about.
It is unwise to leave suspicous-looking code unaddressed.
Concrete issues, like #11864, should also be addressed.

Modification:
The Recycler has been rewritten to use a much simpler algorithm, and now completely avoids the use of weak references.

The WeakOrderQueue is gone, and there is no longer a concept of a shared max capacity - only thread-local capacity.

There is also no more scavenging.
It turns out that other code was relying on the ability to forget references to objects they acquired from a Recycler, and this is hard to support if there is a shared pool for moving handles between threads.

Each thread now has a local pool (an MPSC queue).
Calls to get() and recycle() in the same thread will just operate on the local pool queue.
Calls to recycle() from a thread that did not get() that object (the object was acquired in another thread) will rely on the multi-producer safety of the queue to directly release it back to the local pool that the object came from.
We rely on setting an upper bound for the queue sizes, and the fallible offer() method, for restricting the max thread-local capacity of the recycler.

The constructors have been modified to deprecate the many configurations that are no longer relevant.
A new configuration parameter, the queue chunk-size, is exposed as a constructor parameter.
The queues will grow and shrink in units of this chunk size.

The benchmark from #8052 is copied over and updated, and shows that these changes help overall, with no hidden surprises.
Performance is roughly the same, and allocation rate is down a notch in all the benchmarks.

Result:
The Recycler code is much easier to understand.