Eagerly clear thread-local recycler pools
Motivation:
An object acquired in thread A but strongly referenced from thread B, will keep the entire thread-local pool for A alive past the termination of A.

Modification:
Add an onRemoval callback to the thread-local that holds the local pool instances.
A call to FastThreadLocal.remove() will then clear the queue, reducing the memory pressure from objects strongly held in other threads.

Result:
Reduced memory overhead when recyclable objects are held past the life-time of the thread-local pool they were allocated from.