Remove SCTP support
Motivation:
There is little usage of this protocol, and not all features are supported.
We would rather spend our effort on having excellent TCP, UDP, and domain socket support.

Modification:
The SCTP channels, configs, and related implementation are removed with no replacement.
If need be, the code can be resurrected as a contrib project.

Result:
We no longer support SCTP
Fixes #11555