Fix testsuite which was broken due ba3a58e2efc14e6754dfd3d7c8a8b59f91c7f87e
Motivation:

ba3a58e2efc14e6754dfd3d7c8a8b59f91c7f87e intoduced a change which guards against loading native libraries if there are multiple of the same on the classpath. Unfortunally this also broke our native testsuites as the testsuites did include the native libs two times on the classpath.

Modifications:

Add an exclude which ensures we not copy the native libs to the test-classes folder and so end up with duplicates on the classpath

Result:

Testsuite passes again