JdkZlibEncoder uses pooled heap buffers for deflater input
Motivation:

Previously, when the input ByteBuf was not heap allocated we copied the
input into a new byte array allocated on the heap for every encode() call.
This puts high-throughput clients that use the JdkZlibEncoder under memory
pressure.

Modifications:

Now, when the input ByteBuf is not heap allocated we copy the input into a
heap ByteBuf allocated from the PooledByteBufAllocator, releasing it after
the encode() has completed.

Result:

The result is less heap allocation and GC activity.

Fixes #11890

Signed-off-by: Daniel Ferstay <dferstay@splunk.com>