Improve support for new-api Buffer writes in transports
Motivation:
The transports have previously been relying on the ByteBufConvertible interface for being able to
write the new Buffers to channels.
When ByteBuf is eventually removed, so will that conversion interface.

Modification:
Add first-class support for _writing_ `io.netty.buffer.api.Buffer` instances to channels.
The most important changes are in the ChannelOutboundBuffer, which now has proper support for Buffer.
The transports then rely on the ChannelOutboundBuffer for producing ByteBuffers that are then written to channels.
Test coverage has also been expanded to include Buffer writes in many places.
A number of ByteBuf-specific utility classes have also gotten Buffer-specific counterparts.

Result:
Conversion of the transports to have first-class support for Buffer is far from complete, but this is an important step in the right direction.