Allow same native libraries if content is the same
Motivation:

Sometimes a user might end up with the same native library on the classpath, for example if the user depends on the uber-jar but another dependency directly depend on the classifier version. As long as the native libraries are the same there is no reason to fail loading it.

Modification:

- Check if the native library is the same (by compare the SHA256 of the file) and if so not fail
- Add a unit test

Result:

More robust check for native libraries