Fix ArrayIndexOutOfBounds: 39 from PoolArena.findSubpagePoolHead
Motivation:
We were getting occasional reports of an ArrayIndexOutOfBounds from PoolArena.findSubpagePoolHead, which always happened at index 39.
The sub-page size index array is 39 entries in length, so this access was one past the end.
This should not happen, because a size that *would* compute to index 39, would be a normal-sized allocation (rather than a small one) and thus not use sub-pages at all.
The reason this occurred, is because the "handle" associated with the allocation indicated that it was small, and thus upon freeing we would access the sub-page array, but the specific index was computed from the normalized allocation size, which was "normal" instead of "small".
This happened because of a use-after-free issue in PoolThreadCache, where the entire state of Entry objects were captured prior to return the entry to the recycler *except* for the normCapacity field.
The normCapacity field is what becomes the normSize down the stack, and accessing this field after a recycle could cause the access to race with a concurrent allocation.
So if the allocation was originally small and using sub-pages, then we would capture a sub-page handle, then recycle the entry object, which then got used in a normal-sized allocation, and *then* we read the normCapacity field of this other allocation.

Modification:
The fix is to capture the normCapacity field before calling Entry.recycle() in PoolThreadCache.freeEntry.
A couple of assertions have also been added, to possibly capture bugs like this at an earlier point in the execution.

Result:
We should no longer get this particular ArrayIndexOutOfBoundsException.